Terminal works on MQTT :heavy_check_mark:
